+++ lib/src/extensions.dart
@@ -5,11 +5,7 @@
 // license that can be found in the LICENSE file or at
 // https://opensource.org/licenses/MIT.
 
-import 'package:localregex/localregex.dart';
-import 'package:localregex/src/format_type.dart';
-import 'package:localregex/src/localregex.dart';
-import 'package:localregex/src/localregex_exception.dart';
-import 'package:localregex/src/utils.dart';
+import "../localregex.dart";
 
 /// RegexExtensionForStrings
 ///
@@ -18,7 +14,7 @@ extension RegexExtensionForStrings on String {
   /// clean
   ///
   /// Removes all non-numeric characters from the string.
-  String get clean => replaceAll(RegExp(r'\s+'), '');
+  String get clean => replaceAll(RegExp(r"\s+"), "");
 
   /// lastCharacters
   ///
@@ -45,7 +41,7 @@ extension RegexExtensionForNumbers on String {
         number = regularFormat(this);
       }
     } else {
-      throw LocalRegexException('Mobile number is not valid');
+      throw LocalRegexException("Mobile number is not valid");
     }
 
     return number;
@@ -63,26 +59,26 @@ extension RegexExtensionForID on String {
   String? formatID({IdFormatType formatType = IdFormatType.proper}) {
     late String id;
     if (LocalRegex.isZimID(this)) {
-      final idWithNoSpaces = clean.replaceAll(RegExp('[^A-Z0-9]'), '');
+      final idWithNoSpaces = clean.replaceAll(RegExp("[^A-Z0-9]"), "");
       final idPrefix = idWithNoSpaces.substring(0, 2);
       final idSuffix = idWithNoSpaces.lastCharacters(3);
       final properIdSuffix =
-          '${idSuffix.split('')[0]} ${idSuffix.split('')[1]}${idSuffix.split('')[2]}';
+          "${idSuffix.split("")[0]} ${idSuffix.split("")[1]}${idSuffix.split("")[2]}";
       final idBody = idWithNoSpaces
-          .replaceAll(idPrefix, '')
-          .replaceAll(idSuffix, '')
+          .replaceAll(idPrefix, "")
+          .replaceAll(idSuffix, "")
           .trim();
 
       switch (formatType) {
         case IdFormatType.proper:
-          id = '$idPrefix-$idBody $properIdSuffix';
+          id = "$idPrefix-$idBody $properIdSuffix";
           break;
         case IdFormatType.noSpace:
-          id = '$idPrefix$idBody$properIdSuffix';
+          id = "$idPrefix$idBody$properIdSuffix".clean;
           break;
       }
     } else {
-      throw LocalRegexException('National ID is not valid');
+      throw LocalRegexException("National ID is not valid");
     }
 
     return id;